+++ app/features/tupoksi/TupoksiTahfidzContainer.vue
@@ -1,0 +1,191 @@
+<template>
+  <UCard class="p-4">
+    <div>
+      <h1 class="mb-4 text-center text-2xl font-bold">
+        URAIAN TUGAS AMAL SHOLIH DAN CONTOH KEGIATAN
+      </h1>
+      <table>
+        <thead>
+          <tr>
+            <td class="w-[50%] text-center font-semibold">
+              URAIAN TUGAS AMAL SHOLIH
+            </td>
+            <td class="w-[50%] text-center font-semibold">CONTOH KEGIATAN</td>
+          </tr>
+        </thead>
+        <tbody>
+          <tr>
+            <td>
+              <ul class="list-decimal">
+                <li>
+                  Mendata generus yang punya minat dan kemampuan untuk menghafal
+                  al-Qur&apos;an
+                </li>
+              </ul>
+            </td>
+            <td>
+              <ul class="list-disc">
+                <li>
+                  Membuat formulir pendataan generus yang minat dan mampu
+                  menghafal al-Qur&apos;an.
+                </li>
+                <li>
+                  Jika peserta tahfidz Tingkat kelompok sangat sedikit maka
+                  pelaksanaannya ditempatkan di tingkat desa.
+                </li>
+              </ul>
+            </td>
+          </tr>
+          <tr>
+            <td>
+              <ol class="list-decimal" start="2">
+                <li>
+                  Menyusun program pembinaan tahfidz al-Qur`an untuk menampung
+                  generus yang minat dan punya kemampuan dengan mengacu pada
+                  kurikulum tahfidz
+                </li>
+              </ol>
+            </td>
+            <td>
+              <ul class="list-disc">
+                <li>
+                  Merencanakan waktu khusus kegiatan tahfidz dengan tanpa
+                  mengurangi waktu KBM, diupayakan 3 kali dalam 1 minggu
+                  (contoh: Senin setelah Subuh, Rabu setelah Maghrib, dan Jumat
+                  setelah Maghrib), lalu menentukan jadwal kegiatannya
+                  (menghafal, menyetorkan hafalan, mengulang hafalan).
+                </li>
+                <li>
+                  Menyusun target hafalan setiap setoran hafalan dengan beberapa
+                  kriteria (contoh: ¼ halaman, ⅓ halaman, ½ halaman, 1 halaman),
+                  dan target hafalan disusun dengan urutan mulai juz 30, lalu
+                  juz 1 dan seterusnya.
+                </li>
+                <li>
+                  Membuat jadwal peramutan hafalan bagi para
+                  muballigh/muballighot dan generus usia mandiri yang telah
+                  hafal beberapa surat al- Qur&apos;an.
+                </li>
+                <li>
+                  Membuat jadwal ujian hafalan secara berkala, contoh: 3 bulan
+                  sekali di tingkat kelompok, 6 bulan sekali di tingkat desa dan
+                  1 tahun sekali di tingkat daerah.
+                </li>
+                <li>
+                  Membuat jadwal lomba hafalan bagi generus usia caberawit.
+                </li>
+                <li>
+                  Membuat jadwal pengajian motivasi tahfidz sedaerah setahun
+                  sekali dengan acara motivasi dari PPG bid. tahfidz lalu
+                  dilanjutkan menderes bersama.
+                </li>
+                <li>
+                  Membuat jadwal nasihat pemantapan dari keimaman
+                  kelompok/desa/daerah pada acara pengajian tahfidv.
+                </li>
+              </ul>
+            </td>
+          </tr>
+          <tr>
+            <td>
+              <ol class="list-decimal" start="3">
+                <li>
+                  Melakukan sosialisasi program tahfidz al-Qur'an kepada orang
+                  tua dan pengajar
+                </li>
+              </ol>
+            </td>
+            <td>
+              <ul class="list-disc">
+                <li>
+                  Menyampaikan program tahfidz pada para muballigh/muballighot
+                  saat pengajian mubaligh sedaerah.
+                </li>
+                <li>
+                  Menyampaikan program tahfidz pada musyawaroh Lima Unsur.
+                </li>
+                <li>
+                  Menyampaikan dalil-dalil kefadholan membaca, menghafal dan
+                  mengamalkan al Qur&apos;an.
+                </li>
+                <li>
+                  Menampilkan generus tahfidz di acara pengajian kelompok, desa
+                  dan daerah.
+                </li>
+              </ul>
+            </td>
+          </tr>
+          <tr>
+            <td>
+              <ol class="list-decimal" start="4">
+                <li>
+                  Memberikan bimbingan kepada guru dan generus yang mengikuti
+                  program tahfidz agar berhasil dengan optimal.
+                </li>
+              </ol>
+            </td>
+            <td>
+              <ul class="list-disc">
+                <li>
+                  Memberi pembekalan pada para pengajar tentang SOP dalam KBM
+                  tahfidz:
+                  <ul class="list-[lower-alpha]">
+                    <li>
+                      Mengawali KBM dengan nasihat pemantapan bab jamaah dan
+                      lima bab.
+                    </li>
+                    <li>
+                      Ayat yang akan dihafal dimangkulkan terlebih dahulu bacaan
+                      makna dan keterangannya, adapun cara penyampaian makna dan
+                      keterangan sesuai dengan jenjang usia generus.
+                    </li>
+                    <li>
+                      Target hafalan disesuaikan dengan kemampuan jenjang usia
+                      generus.
+                    </li>
+                    <li>
+                      Mengakhiri kegiatan dengan memberi penjelasan makna yang
+                      terkandung dari ayat yang dihafal, penjelasan mengenai
+                      manfaat menghafal al-Qur&apos;an, dll.
+                    </li>
+                  </ul>
+                </li>
+              </ul>
+            </td>
+          </tr>
+          <tr>
+            <td>
+              <ol class="list-decimal" start="5">
+                <li>
+                  Melakukan monitoring dan evaluasi (monev) terkait pelaksanaan
+                  program tahfidz secara berkala.
+                </li>
+              </ol>
+            </td>
+            <td>
+              <ul class="list-disc">
+                <li>
+                  Turba ke kelompok secara berkala untuk evaluasi tentang
+                  kelancaran KBM tahfidz di kelompok, desa dan daerah serta
+                  membahas solusi dari kendala yang disampaikan.
+                </li>
+              </ul>
+            </td>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+  </UCard>
+</template>
+
+<style scoped>
+  td {
+    border-width: 1px;
+    padding: 0.5rem;
+    padding-left: 1rem;
+  }
+  ul,
+  ol {
+    margin-left: 1rem;
+  }
+</style>
