+++ app/pages/dashboard/(Tahfidz)/tahfidz-tupoksi.vue
@@ -1,0 +1,14 @@
+<script setup lang="ts">
+  import TupoksiTahfidzContainer from "~/features/tupoksi/TupoksiTahfidzContainer.vue";
+  import { useConstantStore } from "~/stores/constant";
+
+  const constantStore = useConstantStore();
+  constantStore.setTitle("Tahfidz / Tupoksi");
+</script>
+
+<template>
+  <Title>Tahfidz | Tupoksi</Title>
+  <main>
+    <TupoksiTahfidzContainer />
+  </main>
+</template>
